Forcing referenceURLs to be http(s) URIs.
This was proposed in VOResource-1_1-Next; the rationale:

Various data providers get the reference URLs wrong (e.g., using plain domain
names) or fail to give sensible values and are never reminded to put them in.
The schema does not notice that because xs:anyURI is rather lenient (by
necessity).

On the other hand, the having something a browser can in all likelihood deal
with appears eminently useful. While that won't prevent 404s, just forcing
referenceURL to start with https?:// would go a long way to ensure useful
reference URLs.